requirements-test.txt
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
python-pptx>=0.6.23
